requirements.txt
pandas==1.5.3
python-dotenv==0.19.2
openpyxl==3.1.2